Add MainComponent and re-structure app
Create a MainComponent that populates instruction-list data based
on the UploadService response. This is the normal usage. This
requires that instruction-list takes input arguments of filename
and a list of instructions instead of being hardcoded to use
the upload service.

This prepares the way for display of arbitrary instruction traces
regardless of their source.